Fix Linux build: move Gate out of the Apple-gated test file
The simplify pass moved the Gate latch into TestTransport.swift, which is
wrapped in #if os(macOS) || os(iOS) || os(tvOS) || os(watchOS), but
AuthenticationTests.swift is not platform-gated and uses Gate — so Linux
failed to compile while macOS passed.

Gate now lives in its own ungated file. Verified no other symbol defined
in a platform-gated test file is referenced from an ungated one.
